fixup! transformer: emit CycloneDX dependency graph
